modify: move `base_ui.py` into `dev` folder
