+++ cogs/wish.py
@@ -20,7 +20,7 @@ from apps.genshin import (
 )
 from apps.genshin.checks import check_cookie
 from apps.text_map import text_map, to_ambr_top
-from base_ui import capture_exception
+from dev.base_ui import capture_exception
 from data.game.standard_characters import get_standard_characters
 from dev.exceptions import FeatureDisabled
 from ui.wish import ChooseBanner, ChooseWeapon, SetAuthKey, WishFilter
