+++ ui/genshin/Abyss.py
@@ -8,7 +8,7 @@ from apps.db import get_user_lang, get_user_theme
 from apps.draw import main_funcs
 from apps.draw.main_funcs import draw_abyss_one_page
 from apps.text_map import text_map
-from base_ui import BaseView
+from dev.base_ui import BaseView
 from dev.models import AbyssResult, CustomInteraction, DefaultEmbed, DrawInput
