+++ ui/genshin/AbyssEnemy.py
@@ -9,7 +9,7 @@ from apps.db import get_user_theme
 from apps.draw import main_funcs
 from apps.draw.utility import image_gen_transition
 from apps.text_map import text_map, to_ambr_top
-from base_ui import BaseView
+from dev.base_ui import BaseView
 from dev.models import AbyssHalf, CustomInteraction, DefaultEmbed, DrawInput
 from utility import divide_chunks
