+++ ui/genshin/DailyReward.py
@@ -10,7 +10,7 @@ import config
 from apps.db import get_user_lang
 from apps.genshin import GenshinApp
 from apps.text_map import text_map
-from base_ui import BaseView
+from dev.base_ui import BaseView
 from dev.models import CustomInteraction, DefaultEmbed, ErrorEmbed
 from utility import divide_chunks, get_dt_now
