+++ ui/genshin/Domain.py
@@ -12,7 +12,7 @@ from apps.db import get_user_lang, get_user_theme
 from apps.draw import main_funcs
 from apps.genshin import get_farm_data, get_uid, get_uid_tz
 from apps.text_map import text_map
-from base_ui import BaseView
+from dev.base_ui import BaseView
 from dev.models import DefaultEmbed, DrawInput
 from utility import get_dt_now
