+++ ui/genshin/EnkaProfile.py
@@ -10,7 +10,7 @@ import dev.models as models
 from apps.db import get_user_theme
 from apps.draw.main_funcs import draw_artifact_card
 from apps.text_map import text_map
-from base_ui import EnkaView
+from dev.base_ui import EnkaView
 from dev.exceptions import FeatureDisabled
 from ui.genshin import EnkaDamageCalc, ProfileSettings
 from ui.others.settings import CustomImage
