+++ ui/genshin/ProfileSettings.py
@@ -7,7 +7,7 @@ import config
 from apps.db.utility import create_user_settings, get_profile_ver
 from apps.genshin import edit_enka_cache, get_uid
 from apps.text_map import text_map
-from base_ui import BaseView, EnkaView
+from dev.base_ui import BaseView, EnkaView
 from dev.exceptions import UIDNotFound
 from dev.models import CustomInteraction, DefaultEmbed, ErrorEmbed
 from ui.genshin.EnkaDamageCalc import GoBack
