+++ ui/genshin/ReminderMenu.py
@@ -12,7 +12,7 @@ from apps.genshin import (
     get_weapon_emoji,
 )
 from apps.text_map import text_map, to_ambr_top
-from base_ui import BaseModal, BaseView
+from dev.base_ui import BaseModal, BaseView
 from dev.exceptions import NumbersOnly
 from dev.models import DefaultEmbed
 from ui.genshin import TalentNotificationMenu, WeaponNotificationMenu
