+++ ui/others/Feedback.py
@@ -3,7 +3,7 @@ from discord.ui import TextInput
 
 import config
 from apps.text_map import text_map
-from base_ui import BaseModal
+from dev.base_ui import BaseModal
 from dev.models import DefaultEmbed
