+++ ui/others/ManageAccounts.py
@@ -12,7 +12,7 @@ import config
 from apps.db import get_user_lang
 from apps.genshin import get_account_select_options, get_uid_region_hash
 from apps.text_map import text_map, to_hutao_login_lang
-from base_ui import BaseModal, BaseView
+from dev.base_ui import BaseModal, BaseView
 from cogs.login import register_user
 from dev.models import CustomInteraction, DefaultEmbed, ShenheBot
 from utility import log
