+++ ui/others/SettingsMenu.py
@@ -7,7 +7,7 @@ import dev.asset as asset
 import config
 from apps.db import get_user_auto_redeem, get_user_lang, get_user_theme
 from apps.text_map import text_map
-from base_ui import BaseView, GoBackButton
+from dev.base_ui import BaseView, GoBackButton
 from data.others.language_options import lang_options
 from dev.models import CustomInteraction, DefaultEmbed, OriginalInfo
 from ui.others.settings import CustomImage, Notif
