+++ ui/others/settings/CustomImage.py
@@ -9,7 +9,7 @@ import config
 from ambr import AmbrTopAPI
 from apps.genshin import get_character_emoji
 from apps.text_map import text_map, to_ambr_top
-from base_ui import BaseModal, BaseView, EnkaView
+from dev.base_ui import BaseModal, BaseView, EnkaView
 from data.game.elements import get_element_emoji, get_element_list
 from dev.models import CustomInteraction, DefaultEmbed, ErrorEmbed
 from ui.genshin import EnkaDamageCalc
