+++ ui/todo/TodoList.py
@@ -11,7 +11,7 @@ from ambr import AmbrTopAPI, Material
 from apps.db import get_user_lang, get_user_theme
 from apps.draw import main_funcs
 from apps.text_map import text_map, to_ambr_top
-from base_ui import BaseModal, BaseView
+from dev.base_ui import BaseModal, BaseView
 from utility.todo_paginator import TodoPaginator, TodoPaginatorView
