+++ utility/paginator.py
@@ -6,7 +6,7 @@ from discord import ui
 import config
 from apps.db import get_user_lang
 from apps.text_map import text_map
-from base_ui import BaseView
+from dev.base_ui import BaseView
 from dev.models import CustomInteraction
